assignments/2/CC_HW2_9931099/step2/requirements.txt
redis 
flask
flask_cors
redis
requests
python-dotenv